Updated tests. Updated names to be idiomatic.
