+++ clients/interfaces/url.go
@@ -14,13 +14,13 @@
 
 package interfaces
 
-// ClientURL is the interface for types that need to define some way to retrieve ClientURL information about services.
+// URLClient is the interface for types that need to define some way to retrieve URLClient information about services.
 // This information can be anything that must be determined at runtime, whether it is unknown or simply not yet known.
-type ClientURL interface {
-	// URLPrefix returns the ClientURL base path (or root) of a service.
+type URLClient interface {
+	// Prefix returns the URLClient base path (or root) of a service.
 	// This is the common root of all REST calls to the service,
 	// and is defined on a per service (rather than per endpoint) basis.
-	// URLPrefix returns the root ClientURL for REST calls to the service if it was able to retrieve that ClientURL;
+	// Prefix returns the root URLClient for REST calls to the service if it was able to retrieve that URLClient;
 	// it returns an error otherwise.
-	URLPrefix() (string, error)
+	Prefix() (string, error)
 }
