+++ clients/urlclient/factory.go
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2019 Dell Inc.
+ * Copyright 2020 Dell Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -12,25 +12,19 @@
  * the License.
  *******************************************************************************/
 
-package rest
+// urlclient provides concrete implementation types that implement the URLClient interface.
+// These types should all, in some way or another, provide some mechanism to fill in service data at runtime.
+package urlclient
 
 import (
+	"github.com/edgexfoundry/go-mod-core-contracts/clients/interfaces"
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/types"
 )
 
-// localClient defines a ClientURL implementation that returns the struct field for the URL.
-type localClient struct {
-	url string
-}
-
-// newLocalClient returns a pointer to a localClient.
-func newLocalClient(params types.EndpointParams) *localClient {
-	return &localClient{
-		url: params.Url,
+// New provides the correct concrete implementation of the URLClient given the params provided.
+func New(params types.EndpointParams, m interfaces.Endpointer) interfaces.URLClient {
+	if params.UseRegistry {
+		return newRegistryClient(params, m, 10)
 	}
-}
-
-// URLPrefix always returns the URL statically defined on object creation.
-func (c *localClient) URLPrefix() (string, error) {
-	return c.url, nil
+	return newLocalClient(params)
 }
