+++ clients/urlclient/registry.go
@@ -12,7 +12,7 @@
  * the License.
  *******************************************************************************/
 
-package rest
+package urlclient
 
 import (
 	"errors"
@@ -22,7 +22,7 @@ import (
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/types"
 )
 
-// registryClient defines a ClientURL implementation that checks for an update from an asynchronously run
+// registryClient defines a URLClient implementation that checks for an update from an asynchronously run
 // EndpointMonitor that will emit a correct URL from the remote registry.
 type registryClient struct {
 	url         string
@@ -33,7 +33,7 @@ type registryClient struct {
 var TimeoutError = errors.New("unable to initialize client")
 
 // newRegistryClient returns a pointer to a registryClient.
-// A pointer is used so that when using configuration from a registry, the URLPrefix can be updated asynchronously.
+// A pointer is used so that when using configuration from a registry, the Prefix can be updated asynchronously.
 func newRegistryClient(params types.EndpointParams, m interfaces.Endpointer, timeout int) *registryClient {
 	e := registryClient{
 		timeout:     timeout,
@@ -53,9 +53,9 @@ func newRegistryClient(params types.EndpointParams, m interfaces.Endpointer, tim
 	return &e
 }
 
-// URLPrefix waits for ClientURL to be updated for timeout seconds. If a value is loaded in that time, it returns it.
+// Prefix waits for URLClient to be updated for timeout seconds. If a value is loaded in that time, it returns it.
 // Otherwise, it returns an error.
-func (c *registryClient) URLPrefix() (string, error) {
+func (c *registryClient) Prefix() (string, error) {
 	if c.initialized {
 		return c.url, nil
 	}
